+++ editors/emacs/lambdapi-proofs.el
@@ -180,12 +180,17 @@ This works for both graphical and text displays."
                     (lp--draw-horizontal-line)
                     (insert gstr))
                   goalsstr))
-        (erase-buffer))
+        (remove-overlays)
+        (erase-buffer)
+        (insert "No goals"))
       (read-only-mode 1))))
 
 (defun lp-display-logs (logs)
   "Display logs in *lp-logs* buffer"
   (let ((logbuf (get-buffer-create "*lp-logs*")))
+    (setq logs (replace-regexp-in-string
+                "^[ \t\n\r]*\\(\\(\u001b\\[[0-9]+m\\)*\\)[ \t\n\r]*" "\\1"
+                logs))
     (with-current-buffer logbuf
       (read-only-mode +1)
       (with-silent-modifications
@@ -286,14 +291,32 @@ This works for both graphical and text displays."
   "Proves till the command/tactic at cursor"
   (interactive)
   (save-excursion
-    (let ((line-empty ; line empty or is a single line comment
-           (save-excursion
-             (beginning-of-line)
-             (looking-at-p "\\([[:space:]]*\\|//.*\\)$"))))
-      (lp-prove-till
-       (if line-empty
-           (lp--prev-command-pos (1+ (point)))
-         (lp--next-command-pos (1- (point))))))))
+    (if (and (not (lp--in-comment-p))
+             (save-excursion (beginning-of-line)
+                             (looking-at-p "^[[:space:]]*$")))
+        (progn
+          (end-of-line)
+          (lp-prove-till (point)))
+      (if (string-match "[\n\t ]" (string (char-before (point))))
+          (re-search-forward "[^\n\t ]"))
+      (if (save-excursion
+            (skip-syntax-forward ".")
+            (lp--in-comment-p))
+          (progn
+            (skip-syntax-forward ".")
+            (let ((comment-type (nth 7 (syntax-ppss))))
+              (goto-char (nth 8 (syntax-ppss)))
+              (forward-comment 1)
+              (skip-syntax-backward "> ")
+              (lp-prove-till (1- (point)))))
+        (let ((line-empty ; line empty or is a single line comment
+               (save-excursion
+                 (beginning-of-line)
+                 (looking-at-p "\\([[:space:]]*\\|//.*\\)$"))))
+          (lp-prove-till
+           (if line-empty
+               (lp--prev-command-pos (1+ (point)))
+             (lp--next-command-pos (1- (point))))))))))
 
 (defun lp-prove-till (pos)
   "Evaluates till pos and moves the cursor to the end of evaluated region"
@@ -359,7 +382,6 @@ and 0 if there is no previous command"
   "Returns the position of the next command's terminator
 and pos if there is no next command"
   (setq npos (1+ (or pos (point))))
-  
   (save-excursion
     (let ((term-regex
            (mapconcat